chat/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["minkj1992 <minkj1992@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
langchain = "^0.0.285"
jinja2 = "^3.1.2"
websockets = "^11.0.3"
loguru = "^0.7.1"
httpx = "^0.24.1"
openai = "^0.28.0"
redis-om = "^0.2.1"
redis = "^4.0.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
html2text = "^2020.1.16"
tiktoken = "^0.4.0"
transformers = {extras = ["torch"], version = "^4.33.1"}
huggingface-hub = "^0.17.1"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.21.1"
pytest = "^7.4.2"
asgi-lifespan = "^2.1.0"
pytest-mock = "^3.11.1"
locust = "^2.16.1"
websocket-client = "^1.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
  "."
]

